Make all the things work - a little bit
This is the base setup code for replacing sidekiq with
headbutt. It has some basic tests that use the bunny-mock
testing library around the Processor object which
is responsible processing messages off the rabbitMQ
queue.

At this point the threading, error recover code is not
tested, it has mostly been copied from sidekiq with
some modification to (hopefully) make it easier to 
test later.

Issues: The retry code is aware of the backend 
implementation. I would prefer that this could be 
avoided as this have allowed the a plugin to be written 
that would use a reds backend. :)